feat(context): migrated shop state to Context API, moved logic from Layout.jsx to ShopContext.js
