+++ frontend-fiasco/src/context/ShopContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const ShopContext = createContext();
