+++ frontend-fiasco/src/main.jsx
@@ -3,11 +3,14 @@ import { createRoot } from 'react-dom/client'
 import "./styles/index.css";
 import routes from './routes/routes.jsx'
 import { createBrowserRouter, RouterProvider} from 'react-router-dom'
+import { ShopProvider } from './context/ShopContext.jsx';
 
 const router = createBrowserRouter(routes);
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <RouterProvider router={router}/>
+    <ShopProvider>
+        <RouterProvider router={router}/>
+      </ShopProvider>
   </StrictMode>,
 )
