+++ frontend-fiasco/src/pages/Cart.jsx
@@ -1,11 +1,13 @@
 import ProductItem from "../components/ProductItem";
-import { Link, useOutletContext } from "react-router-dom";
+import { Link } from "react-router-dom";
 import CartItem from "../components/CartItem";
 import CartMessage from "../components/CartMessage";
+import { useContext } from "react";
+import { ShopContext } from "../context/ShopContext";
 
 function Cart() {
 
-    const { cart, removeFromCart, calculateCartTotal, calculateItemsInCart} = useOutletContext();
+    const { cart, removeFromCart, calculateCartTotal, calculateItemsInCart } = useContext(ShopContext);
 
     return (
         <>
@@ -22,15 +24,17 @@ function Cart() {
                 ))}
             </ul>
             
-        <div className="cart-summary">
-            <h3>Cart Summary</h3>
-            <p>Items in Cart: {calculateItemsInCart()}</p>
-            <p>Subtotal: ${calculateCartTotal()}</p>
-            <p>Taxes (10%): ${ (calculateCartTotal() * 0.1).toFixed(2) }</p>
-            <p><strong>Grand Total: ${(calculateCartTotal() * 1.1).toFixed(2)}</strong></p>
-            <button>Checkout</button>
-            <Link to="/products"><button>Continue Shopping</button></Link>
-        </div>
+            <div className="cart-summary">
+                <h3>Cart Summary</h3>
+                <p>Items in Cart: {calculateItemsInCart()}</p>
+                <p>Subtotal: ${calculateCartTotal()}</p>
+                <p>Taxes (10%): ${ (calculateCartTotal() * 0.1).toFixed(2) }</p>
+                <p><strong>Grand Total: ${(calculateCartTotal() * 1.1).toFixed(2)}</strong></p>
+                <button>Checkout</button>
+                <Link to="/products">
+                    <button>Continue Shopping</button>
+                </Link>
+            </div>
 
         </div>
         </>
