+++ frontend-fiasco/src/utils/api.js
@@ -2,8 +2,6 @@ async function  fetchProducts() {
     try {
         const response =  await fetch('https://fakestoreapi.com/products')
         const data = await response.json()
-        // const updatedData = data.map(product => ({ ...product, quantity: 1 })); // Adding quantity field
-        // return updatedData
         return data
     }
      catch (error) {
